+++ src/frontend.cpp
@@ -2186,7 +2186,7 @@ MenuNumber create_menu(struct GuiMenu *gmnu)
     }
     // Make scale factor
     int units_per_px;
-    units_per_px = min(units_per_pixel,units_per_pixel_min*16/10);
+    units_per_px = min((int)units_per_pixel,units_per_pixel_min*16/10);
     // Decrease scale factor if for some reason resulting size would exceed screen (wierd aspec ratio support)
     if (gmnu->width * units_per_px > LbScreenWidth() * 16)
         units_per_px = LbScreenWidth() * 16 / gmnu->width;
@@ -3351,7 +3351,7 @@ void frontend_update(short *finish_menu)
     case FeSt_MAIN_MENU:
         frontend_button_info[8].font_index = (continue_game_option_available?1:3);
         //this uses original timing function for compatibility with frontend_set_state()
-        if ( abs(LbTimerClock()-time_last_played_demo) > MNU_DEMO_IDLE_TIME )
+        if ( abs(LbTimerClock()-(long)time_last_played_demo) > MNU_DEMO_IDLE_TIME )
           frontend_set_state(FeSt_DEMO);
         break;
     case FeSt_FELOAD_GAME:
